reorganize the export files
